+++ Frontend/app/src/main/java/fragments/profile/NutritionSectionFragment.java
@@ -1,4 +1,4 @@
-package fragments;
+package fragments.profile;
 
 import android.os.Bundle;
 
@@ -11,10 +11,10 @@ import android.example.weightworks.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link MetricSectionFragment#newInstance} factory method to
+ * Use the {@link NutritionSectionFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class MetricSectionFragment extends Fragment {
+public class NutritionSectionFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -25,7 +25,7 @@ public class MetricSectionFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public MetricSectionFragment() {
+    public NutritionSectionFragment() {
         // Required empty public constructor
     }
 
@@ -35,11 +35,11 @@ public class MetricSectionFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment MetricSectionFragment.
+     * @return A new instance of fragment NutritionSectionFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static MetricSectionFragment newInstance(String param1, String param2) {
-        MetricSectionFragment fragment = new MetricSectionFragment();
+    public static NutritionSectionFragment newInstance(String param1, String param2) {
+        NutritionSectionFragment fragment = new NutritionSectionFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -60,6 +60,6 @@ public class MetricSectionFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_metric_section, container, false);
+        return inflater.inflate(R.layout.fragment_nutrition_section, container, false);
     }
 }
