+++ Frontend/app/src/main/res/layout/fragment_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="fragments.HistoryFragment">
+    tools:context="fragments.history.HistoryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/history_rv"
