+++ Frontend/app/src/main/res/layout/fragment_metric_section.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="fragments.MetricSectionFragment">
+    tools:context="fragments.profile.MetricSectionFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
