+++ Frontend/app/src/main/res/layout/fragment_new_workout.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="fragments.NewWorkoutFragment">
+    tools:context="fragments.workout.NewWorkoutFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/newWorkout_title_text"
