+++ Frontend/app/src/main/res/layout/fragment_workout_draft_section.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="fragments.WorkoutDraftSectionFragment">
+    tools:context="fragments.workout.WorkoutDraftSectionFragment">
 
     <TextView
         android:layout_width="match_parent"
